pylint dupe fix: refactor pixlematrix and matrix_featherwing to use superclass
